requirements.txt
Flask==2.1.0
redis==3.5.3
PyJWT==2.3.1
python-dotenv==0.19.0
